refactor: DoublyLinkedList.cを分割し構造を改善
詳細:
- LinkedList.cをDoublyLinkedList.cとしてリファクタリング。
- コメントとドキュメントコメントを追加し、可読性と保守性を向上。
- 全体のファイル構造を整理。
